Remove system() + mkdir call
Prior to this patch, creating the destination directory was done using a
system() call to mkdir -p. This was done to make sure parent directories
were created. This complicates SELinux rules, and adds at least a little
risk as pstore-clean runs as root.

This patch uses the cutils fs_mkdirs() function which creates the full
path, including parents if necessary.

Change-Id: Ia92848152bfb2c7166ef9f418e757cb4b3d3e9a2
Signed-off-by: Michael Bergeron <michael.bergeron@intel.com>