Build 64bit binaries for recovery and userfastboot on 32p builds
32p builds set the primary ABI to 64bit and secondary ABI to 32bit but
also set a flag to prefer 32bit binaries be built. This works well for
the system image but causes problem for the recovery and userfastboot
images. For these images, we need to use 64bit for all dynamic binaries
and libraries. We use the LOCAL_MULTILIB feature to force the binaries
being used in these images to be 64bit in this case. The 32o image will
continue to have 32bit binaries.

Change-Id: I7cb5e11c14b18ba7fedc730611d7b666b311156b
Signed-off-by: Bruce Beare <bruce.j.beare@intel.com>